Add email password reset and welcome email features
Configured SMTP email backend and moved sensitive credentials to things.py. Added password reset URLs and updated login template to use Django's password reset flow. Implemented welcome email on profile creation and improved user deletion error handling in signals.py. Updated account template to hide 'Send Message' button for own profile.

+++ users/signals.py
@@ -1,6 +1,10 @@
 
 from django.db.models.signals import post_save, post_delete
 from django.dispatch import receiver
+from django.core.mail import send_mail
+from django.conf import settings
+from django.contrib.auth.models import User
+from django.core.exceptions import ObjectDoesNotExist
 
 from django.contrib.auth.models import User
 from .models import Profile
@@ -17,6 +21,16 @@ def createProfile(sender, instance, created, **kwargs):
             email = user.email,
             name = user.first_name,
         )
+        subject = "Welcome to DevSearch!"
+        message = 'We are glad you are here'
+        send_mail(
+            subject,
+            message,
+            settings.EMAIL_HOST_USER,
+            [profile.email],
+            fail_silently=False,
+        )
+
 
 
 def updateUser(sender, instance, created, **kwargs):
@@ -28,9 +42,15 @@ def updateUser(sender, instance, created, **kwargs):
         user.email = profile.email
         user.save()
 
+
+
 def deleteUser(sender, instance, **kwargs):
-    user = instance.user
-    user.delete()
+    try:
+        user = instance.user
+        if user:
+            user.delete()
+    except (User.DoesNotExist, ObjectDoesNotExist):
+        pass
 
 post_save.connect(createProfile, sender=User)
 post_save.connect(updateUser, sender=Profile)
